Renamed Polygon to Polyhedron.
Now, to actually write a test for this Polygon.

There is a cmake test, but it only returns the OK status.
