+++ src/Tests/Utilities/Collision/GJKPolyhedron.cpp
@@ -1,0 +1,11 @@
+#include "../../../Utilities/Collision/GJKPolyhedron.h"
+#include <iostream>
+
+#include "Helper.h"
+
+int main() {
+    const static int FAILURE = 1;
+    const static int SUCCESS = 0;
+
+    return SUCCESS;
+}
